flower_bot/requirements.txt
numpy==1.15.0
Keras==2.2.4
Flask==1.0.2
Werkzeug==0.14.1
Pillow==6.0.0
tensorflow==1.13.1
opencv-python==3.3.0.10
